+++ packages/fetcher/src/error.ts
@@ -1,3 +1,5 @@
+import { RuntimeError } from '@distributedlab/tools'
+
 import type { FetcherRequest, FetcherResponse } from '@/types'
 
 export class FetcherError<T = undefined> extends Error {
@@ -9,5 +11,13 @@ export class FetcherError<T = undefined> extends Error {
     super(resp.statusText)
     this.response = resp
     this.request = resp.request
+  }
+}
+
+export class FetcherURLParseError extends RuntimeError {
+  public name = 'FetcherURLParseError'
+
+  constructor(e = new TypeError('Failed to parse URL')) {
+    super(e)
   }
 }