+++ packages/fetcher/src/helpers/request.test.ts
@@ -1,7 +1,10 @@
-import type { FetcherAbortManager } from '@/abort-manager'
 import { DEFAULT_CONFIG } from '@/const'
 import type { HTTP_METHODS } from '@/enums'
-import type { FetcherConfig, FetcherRequestConfig } from '@/types'
+import type {
+  FetcherAbortManager,
+  FetcherConfig,
+  FetcherRequestConfig,
+} from '@/types'
 
 import { buildRequestConfig, buildRequestURL } from './request'