+++ packages/fetcher/src/types/interceptor.ts
@@ -16,3 +16,15 @@ export type FetcherInterceptor = {
   response?: FetcherResponseInterceptor
   error?: FetcherErrorResponseInterceptor
 }
+
+export interface FetcherInterceptorManager {
+  readonly interceptors: FetcherInterceptor[]
+
+  add(interceptor: FetcherInterceptor): void
+  remove(interceptor: FetcherInterceptor): void
+  clear(): void
+
+  runRequestInterceptors(r: FetcherRequest): Promise<FetcherRequest>
+  runResponseInterceptors<T>(r: FetcherResponse<T>): Promise<FetcherResponse<T>>
+  runErrorInterceptors<T>(r: FetcherResponse<T>): Promise<FetcherResponse<T>>
+}