+++ packages/fetcher/src/types/response-builder.ts
@@ -1,0 +1,6 @@
+import type { FetcherResponse } from './fetcher'
+
+export interface FetcherResponseBuilder<T> {
+  build(): Promise<FetcherResponse<T>>
+  populateResponse(response: Response): FetcherResponseBuilder<T>
+}